+++ src/main/java/br/com/taskcontroller/Record/COMBO/PapelComboDTO.java
@@ -1,8 +1,9 @@
 package br.com.taskcontroller.Record.COMBO;
 
+import java.math.BigDecimal;
+
 public record PapelComboDTO(
         Long idpapel,
         String descpapel,
-        Long valorhora
-) {
+        BigDecimal valorhora) {
 }
